Cross tenant thing protection

+++ app/handlers/thing.handler.go
@@ -18,10 +18,11 @@ func NewThingAPI(thingService services.ThingServiceInterface) *ThingHandler {
 	return &ThingHandler{thing: thingService}
 }
 
-func (handler *ThingHandler) Create(ctx *gin.Context) {
+func (handler *ThingHandler) CreateThing(ctx *gin.Context) {
 	var newThing models.Thing
 	ctx.BindJSON(&newThing)
-	if middleware.IsAuthorizationAtLeast(ctx, "Admin") {
+	organization, _ := middleware.GetOrganizationClaim(ctx)	
+	if middleware.IsAuthorizationAtLeast(ctx, "Admin") && organization.ID == newThing.OrganizationId {
 		err := handler.thing.Create(ctx.Request.Context(), &newThing)
 		if err == nil {
 			result := utils.SuccessPayload(newThing, "Succesfully created thing")
@@ -49,12 +50,17 @@ func (handler *ThingHandler) UpdateThing(ctx *gin.Context) {
 	var thing models.Thing
 	ctx.BindJSON(&thing)
 	if middleware.IsAuthorizationAtLeast(ctx, "Admin") {
-		err := handler.thing.Update(ctx.Request.Context(), &thing)
-		if err != nil {
-			utils.Response(ctx, http.StatusBadRequest, utils.NewHTTPCustomError(utils.BadRequest, err.Error()))
+		organization, _ := middleware.GetOrganizationClaim(ctx)
+		if organization.ID == thing.ID { // Cross-tenant update protection
+			err := handler.thing.Update(ctx.Request.Context(), &thing)
+			if err != nil {
+				utils.Response(ctx, http.StatusBadRequest, utils.NewHTTPCustomError(utils.BadRequest, err.Error()))
+			} else {
+				result := utils.SuccessPayload(nil, "Succesfully updated")
+				utils.Response(ctx, http.StatusOK, result)
+			}
 		} else {
-			result := utils.SuccessPayload(nil, "Succesfully updated")
-			utils.Response(ctx, http.StatusOK, result)
+			utils.Response(ctx, http.StatusUnauthorized, utils.NewHTTPError(utils.Unauthorized))	
 		}
 	} else {
 		utils.Response(ctx, http.StatusUnauthorized, utils.NewHTTPError(utils.Unauthorized))
@@ -63,12 +69,22 @@ func (handler *ThingHandler) UpdateThing(ctx *gin.Context) {
 
 func (handler *ThingHandler) Delete(ctx *gin.Context) {
 	if middleware.IsAuthorizationAtLeast(ctx, "Admin") {
-		err := handler.thing.Delete(ctx.Request.Context(), ctx.Param("thingId"))
+		organization, _ := middleware.GetOrganizationClaim(ctx)
+		thing, err := handler.thing.FindById(ctx, ctx.Param("thingId"))
 		if err != nil {
-			utils.Response(ctx, http.StatusBadRequest, utils.NewHTTPCustomError(utils.BadRequest, err.Error()))
+			if organization.ID == thing.OrganizationId { // Cross-tenant deletion prevention
+				err := handler.thing.Delete(ctx.Request.Context(), ctx.Param("thingId"))
+				if err != nil {
+					utils.Response(ctx, http.StatusBadRequest, utils.NewHTTPCustomError(utils.BadRequest, err.Error()))
+				} else {
+					result := utils.SuccessPayload(nil, "Successfully deleted")
+					utils.Response(ctx, http.StatusOK, result)
+				}
+			} else {
+				utils.Response(ctx, http.StatusUnauthorized, utils.NewHTTPError(utils.Unauthorized))	
+			}
 		} else {
-			result := utils.SuccessPayload(nil, "Successfully deleted")
-			utils.Response(ctx, http.StatusOK, result)
+			utils.Response(ctx, http.StatusBadRequest, utils.NewHTTPCustomError(utils.BadRequest, err.Error()))
 		}
 	} else {
 		utils.Response(ctx, http.StatusUnauthorized, utils.NewHTTPError(utils.Unauthorized))
